Begun Goal/Hazard implementation
Levels now have a coinsRequired field that dictates how many coins are required to pass the level. When zero simply reaching the goal constitutes a pass. In level_tester passing a level simply quits the game.

There is an editor flag for hazards but it has not yet been tested. Also, in main.cpp some code must be added to check if the player is dead and if so, use some alarms or have a screen popup that requires input to "try again", since currently it should just reset the level.