Updated project goals.
This will be my last time "coming back" to this project. My goal is to get it to a playable state with some minimal content, and get a web build running for my site. The priority on shifting to pure C is gone since employers don't care about seeing pure C they want C++, C# or Java.