Player Collision uses OverlapBoxes now
Tried to implement the stair system but I created a bug where moving left or right while ascending past a platform will cause the player to snap to the platform. It might be worthwhile to reinvent the env_object system because having stairs be made of many blocks might not be efficient AND there may be a way to simplify stairs and slopes (I would be happy to only have slopes if it meant there was a way to do it).

Perhaps support for Triangles and other shapes is in order?